schema_graph_critic/requirements.txt
# Core ML Framework
torch>=2.0.0
torch-geometric>=2.4.0

# Graph utilities
networkx>=3.1

# Sentence embeddings for semantic features
sentence-transformers>=2.2.0

# JSON Schema handling
jsonschema>=4.17.0
referencing>=0.30.0

# Data handling
numpy>=1.24.0
pandas>=2.0.0

# Progress bars and utilities
tqdm>=4.65.0
rich>=13.0.0

# Environment variables
python-dotenv>=1.0.0

# Testing
pytest>=7.3.0
